requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
autopep8==1.5.4
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-celery==3.3.1
django-cors-headers==3.5.0
djangorestframework==3.11.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==3.0.37
MarkupSafe==1.1.1
packaging==20.4
prompt-toolkit==3.0.8
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycodestyle==2.6.0
pyparsing==2.4.7
pytz==2020.4
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
